+++ lib/image_optim/bin_resolver/bin.rb
@@ -33,20 +33,34 @@ class ImageOptim
 
       is = ComparableCondition.is
 
-      FAIL_CHECKS = [
-        [:pngcrush, is.between?('1.7.60', '1.7.65'), 'is known to produce '\
-              'broken pngs'],
-        [:pngcrush, is == '1.7.80', 'loses one color in indexed images'],
-        [:pngquant, is < '2.0', 'is not supported'],
-      ].freeze
+      FAIL_CHECKS = {
+        :pngcrush => [
+          [is.between?('1.7.60', '1.7.65'), 'is known to produce broken pngs'],
+          [is == '1.7.80', 'loses one color in indexed images'],
+        ],
+        :pngquant => [
+          [is < '2.0', 'is not supported'],
+        ],
+      }.freeze
 
-      WARN_CHECKS = [
-        [:advpng, is < '1.17', 'does not use zopfli'],
-        [:gifsicle, is < '1.85', 'does not support removing extension blocks'],
-        [:pngcrush, is < '1.7.38', 'does not have blacken flag'],
-        [:pngquant, is < '2.1', 'may be lossy even with quality `100-`'],
-        [:optipng, is < '0.7', 'does not support -strip option'],
-      ].freeze
+      WARN_CHECKS = {
+        :advpng => [
+          [is == 'none', 'is of unknown version'],
+          [is < '1.17', 'does not use zopfli'],
+        ],
+        :gifsicle => [
+          [is < '1.85', 'does not support removing extension blocks'],
+        ],
+        :pngcrush => [
+          [is < '1.7.38', 'does not have blacken flag'],
+        ],
+        :pngquant => [
+          [is < '2.1', 'may be lossy even with quality `100-`'],
+        ],
+        :optipng => [
+          [is < '0.7', 'does not support -strip option'],
+        ],
+      }.freeze
 
       # Fail if version will not work properly
       def check_fail!
@@ -54,8 +68,9 @@ class ImageOptim
           fail UnknownVersion, "could not get version of #{name} at #{path}"
         end
 
-        FAIL_CHECKS.each do |bin_name, matcher, message|
-          next unless bin_name == name
+        return unless FAIL_CHECKS[name]
+
+        FAIL_CHECKS[name].each do |matcher, message|
           next unless matcher.match(version)
 
           fail BadVersion, "#{self} (#{matcher}) #{message}"
@@ -66,11 +81,14 @@ class ImageOptim
       def check!
         check_fail!
 
-        WARN_CHECKS.each do |bin_name, matcher, message|
-          next unless bin_name == name
+        return unless WARN_CHECKS[name]
+
+        WARN_CHECKS[name].each do |matcher, message|
           next unless matcher.match(version)
 
           warn "WARN: #{self} (#{matcher}) #{message}"
+
+          break
         end
       end
 
@@ -85,7 +103,9 @@ class ImageOptim
       # Getting version of bin, will fail for an unknown name
       def version_string
         case name
-        when :advpng, :gifsicle, :jpegoptim, :optipng
+        when :advpng
+          capture("#{escaped_path} --version 2> #{Path::NULL}")[/\bv(\d+(\.\d+)+|none)/, 1]
+        when :gifsicle, :jpegoptim, :optipng
           capture("#{escaped_path} --version 2> #{Path::NULL}")[/\d+(\.\d+)+/]
         when :svgo, :pngquant
           capture("#{escaped_path} --version 2>&1")[/\d+(\.\d+)+/]