+++ CHANGELOG.markdown
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+* Handle `vnone` version of `advpng` [#165](https://github.com/toy/image_optim/issues/165) [@toy](https://github.com/toy)
+
 ## v0.26.2 (2018-08-15)
 
 * Ignore segmentation fault for `pngout` <= `20150920` [#158](https://github.com/toy/image_optim/issues/158) [@toy](https://github.com/toy)